requirements.txt
google-adk
python-dotenv
google-auth
google-cloud-bigquery
uvicorn
fastapi
aiosqlite
